moving logic from plugin.py to odm_theme_helper
