+++ ckanext/odm_theme/templates/scheming/form_snippets/odm_organization.html
@@ -7,7 +7,7 @@
 {% set dataset_has_organization = data.owner_org or data.group_id %}
 {% set organizations_available = h.organizations_available('create_dataset') %}
 {% set user_is_sysadmin = h.check_access('sysadmin') %}
-{% set show_organizations_selector = organizations_available and (user_is_sysadmin or dataset_is_draft or dataset_is_public) %}
+{% set show_organizations_selector = True %}
 {% set show_visibility_selector = dataset_has_organization or (organizations_available and (user_is_sysadmin or dataset_is_draft)) %}
 
 <div data-module="odm_theme_review_system_logic" data-module-owner_org="{{data.owner_org}}">
